made all the papers evenly spaced and responsive
