+++ success-dashboard/src/components/MainDash/GoodVibesNews.js
@@ -3,7 +3,7 @@ import Paper from 'material-ui/Paper';
 
 const style = {
   height: 300,
-  width: 300,
+  width: 'calc(33% - 40px)',
   margin: 20,
   textAlign: 'center',
   display: 'inline-block',
@@ -22,6 +22,7 @@ export default class GoodVibesNews extends React.Component {
     console.log('this props: ', this.props);
     return (
       <Paper style={style} zDepth={5} rounded={false}>
+        <h3>Feel Good News</h3>
       	{this.props.news.map((story, i) => (
           <div>
             <a target='_blank' href={story.data.url}><p>{story.data.title}</p></a>
