+++ success-dashboard/src/components/MainDash/MeetupFeed.js
@@ -3,7 +3,7 @@ import Paper from 'material-ui/Paper';
 
 const style = {
   height: 300,
-  width: 300,
+  width: 'calc(33% - 40px)',
   margin: 20,
   textAlign: 'center',
   display: 'inline-block',
@@ -113,6 +113,3 @@ export default class MeetupFeed extends React.Component {
 // 		)
 // 	}
 // };
-
-
-
