+++ success-dashboard/src/components/MainDash/ToDoList.js
@@ -3,7 +3,7 @@ import Paper from 'material-ui/Paper';
 
 const style = {
   height: 300,
-  width: 500,
+  width: 'calc(50% - 40px)',
   margin: 20,
   textAlign: 'center',
   display: 'inline-block',
@@ -14,4 +14,3 @@ export const ToDoList = () => (
     	<h2>ToDoList</h2>
   	</Paper>
 );
-
