+++ success-dashboard/src/components/MainDash/Trivia.css
@@ -1,0 +1,3 @@
+.yarb {
+  display: flex;
+}
